🚚 Rename from nb-NO to nb
