+++ src/shared/locales/nb/about.js
