+++ src/shared/locales/nb/app.js
