+++ src/shared/locales/nb/edit.js
