+++ src/shared/locales/nb/help.js
