+++ src/shared/locales/nb/index.js
