+++ src/shared/locales/nb/menu.js
