+++ src/shared/locales/nb/preferences.js
