+++ src/shared/locales/nb/subnav.js
