+++ src/shared/locales/nb/task.js
