+++ src/shared/locales/nb/window.js
